+++ application/controllers/Views.php
@@ -1,0 +1,68 @@
+<?php
+
+defined('BASEPATH') OR exit('No direct script access allowed');
+
+class Views extends Application
+{
+
+	/**
+	 * Index Page for this controller.
+	 *
+	 * Maps to the following URL
+	 * 		http://example.com/
+	 * 	- or -
+	 * 		http://example.com/welcome/index
+	 *
+	 * So any other public methods not prefixed with an underscore will
+	 * map to /welcome/<method_name>
+	 * @see https://codeigniter.com/user_guide/general/urls.html
+	 */
+	public function index()
+	{
+		$this->data['pagetitle'] = 'Ordered TODO List';
+		$this->data['pagebody'] = 'template_secondary';
+		$tasks = $this->tasks->all();
+		$this->data['content'] = 'Ok';
+		$this->data['leftside'] = $this->makePrioritizedPanel($tasks);
+		$this->data['rightside'] = $this->makeCategorizedPanel($tasks);
+
+		$this->render('template_secondary'); 
+	}
+
+	function makeCategorizedPanel($tasks)
+	{
+		$parms = ['display_tasks' => $this->tasks->getCategorizedTasks()];
+		return $this->parser->parse('by_category', $parms, true);
+	}
+
+	function makePrioritizedPanel($tasks)
+	{
+		foreach ($tasks as $task)
+		{
+			if ($task->status != 2)
+				$undone[] = $task;
+		}
+		usort($undone, "orderByPriority");
+		foreach($undone as $task)
+		{
+			$task->priority = $this->app->priority($task->priority);
+		}
+		foreach($undone as $task)
+		{
+			$converted[] = (array) $task;
+		}
+		$parms = ['display_tasks' => $converted];
+		return $this->parser->parse('by_priority', $parms, true);
+	}
+}
+
+// return -1, 0, or 1 of $a's priority is higher, equal to, or lower than $b's
+function orderByPriority($a, $b)
+{
+    if ($a->priority > $b->priority)
+        return -1;
+    elseif ($a->priority < $b->priority)
+        return 1;
+    else
+        return 0;
+}